Renamed models/VectorDimensions.py to follow snake case convention
